server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
django==5.0.7
    # via
    #   -r requirements.in
    #   django-ninja
django-ninja==1.2.1
    # via -r requirements.in
pydantic==2.8.2
    # via django-ninja
pydantic-core==2.20.1
    # via pydantic
sqlparse==0.5.0
    # via django
typing-extensions==4.12.2
    # via
    #   asgiref
    #   pydantic
    #   pydantic-core
